ui for syncing indicator and move it to footer
