+++ web/src/components/Footer/Footer.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useRef } from 'react'
+import React, { useState, useRef, useEffect } from 'react'
 import useOnClickOutside from 'use-onclickoutside'
 import { useRouteMatch } from 'react-router-dom'
 import Zoom from '../Zoom/Zoom.connector'
@@ -7,9 +7,12 @@ import Icon from '../Icon/Icon'
 import Button from '../Button/Button'
 
 import MapViewingOptions from '../MapViewingOptions/MapViewingOptions'
-import { CellIdString } from '../../types/shared'
+import { AgentPubKeyB64, CellIdString } from '../../types/shared'
+import { getAdminWs, getAppWs } from '../../hcWebsockets'
+import SyncingIndicator from '../SyncingIndicator/SyncingIndicator'
 
 export type FooterProps = {
+  agentAddress: AgentPubKeyB64
   hiddenAchievedOutcomes: CellIdString[]
   hiddenSmallOutcomes: CellIdString[]
   selectedLayeringAlgo: string
@@ -21,6 +24,7 @@ export type FooterProps = {
 }
 
 const Footer: React.FC<FooterProps> = ({
+  agentAddress,
   hiddenAchievedOutcomes,
   hiddenSmallOutcomes,
   selectedLayeringAlgo,
@@ -63,6 +67,55 @@ const Footer: React.FC<FooterProps> = ({
     }
   }
 
+  // for syncing a project
+  const [dnas, setDnas] = useState([])
+  const [numOpsToFetch, setNumOpsToFetch] = useState(0)
+
+  useEffect(() => {
+    const getDnas = async () => {
+      const adminWs = await getAdminWs()
+      const dnas = await adminWs.listDnas()
+      setDnas(dnas)
+    }
+
+    getDnas()
+  }, [projectId])
+
+  // display syncing indicator when numOpsToFetch > 0
+  useEffect(() => {
+    const fetchOpData = async () => {
+      if (!projectId) {
+        return
+      }
+
+      const activeProjectDna = projectId.substring(0, projectId.indexOf('['))
+      const appWs = await getAppWs()
+      // TODO: check `last_time_queried` parameter to see if its useful
+      const networkInfo = await appWs.networkInfo({
+        agent_pub_key: agentAddress as any,
+        dnas: dnas as any,
+      })
+
+      let i: number
+      let sum = 0
+
+      for (i = 0; i < dnas.length; i++) {
+        const dna = dnas[i].toString()
+
+        if (dna === activeProjectDna)
+          sum = networkInfo[i].fetch_pool_info.num_ops_to_fetch
+      }
+
+      setNumOpsToFetch(sum)
+    }
+
+    const interval = setInterval(() => fetchOpData(), 1000)
+
+    return () => clearInterval(interval)
+  }, [projectId, agentAddress, dnas])
+
+  const isSyncing = numOpsToFetch > 0
+
   return (
     <div className="footer" ref={ref}>
       {/* Report Issue Button */}
@@ -77,6 +130,7 @@ const Footer: React.FC<FooterProps> = ({
       {/* Zooming and Viewing Options on Map View */}
       {projectPage && (
         <div className={bottomRightPanelClassName}>
+          {isSyncing && <SyncingIndicator />}
           {mapPage && (
             <div className="map-viewing-options-button-wrapper">
               {/* If map viewing options is open */}
