+++ web/src/components/SyncingIndicator/SyncingIndicator.scss
@@ -1,0 +1,36 @@
+.syncing-indicator {
+  margin: 10px;
+  position: relative;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+
+  &:hover .tooltip-wrapper {
+    position: absolute;
+    left: 50%;
+    visibility: visible;
+    transition-delay: 0.9s;
+    transition-property: visibility;
+  }
+
+  .icon {
+    animation: spin 2s linear infinite;
+    -webkit-animation: spin 2s linear infinite;
+  }
+}
+
+@keyframes spin {
+  100% {
+    -webkit-transform: rotate(360deg);
+    transform: rotate(360deg);
+  }
+}
+
+@-webkit-keyframes spin {
+  100% {
+    -webkit-transform: rotate(360deg);
+  }
+}
