+++ web/src/routes/App.component.tsx
@@ -261,6 +261,7 @@ const App: React.FC<AppProps> = ({
             {redirToFinishMigration && <Redirect to="/finish-update" />}
             {agentAddress && whoami && (
               <Footer
+                agentAddress={agentAddress}
                 hiddenAchievedOutcomes={hiddenAchievedOutcomes}
                 hiddenSmallOutcomes={hiddenSmallOutcomes}
                 showSmallOutcomes={showSmallOutcomes}
